refactor: Cambia el tipo de dato del campo idBrand y actualiza los test

+++ src/main/java/com/rosales/springtest/dto/PriceRequestDto.java
@@ -22,6 +22,6 @@ public class PriceRequestDto {
 
     @NotNull
     @Min(1)
-    private Integer idBrand;
+    private Long idBrand;
 
 }
