+++ src/main/java/com/rosales/springtest/repository/IPriceRepository.java
@@ -7,6 +7,6 @@ import java.util.Optional;
 
 public interface IPriceRepository extends IGenericRepository<Price, Long> {
     Optional<Price> findFirstByStartDateLessThanEqualAndEndDateGreaterThanEqualAndIdProductAndIdBrandOrderByPriorityDesc(
-            LocalDateTime startDate, LocalDateTime endDate, Long idProduct, Integer idBrand);
+            LocalDateTime startDate, LocalDateTime endDate, Long idProduct, Long idBrand);
 
 }
